fix: signup error

+++ apis/participant.py
@@ -1,8 +1,8 @@
-from flask import Response, jsonify, request
+from flask import Response, jsonify, request, g
 from flask_restx import Namespace, Resource, fields
 
-from model import Participant, Study, db
-
+from model import Participant, Study, db, StudyContributor
+from .login import is_granted
 api = Namespace("Participant", description="Participant operations", path="/")
 
 participant_model = api.model(
@@ -31,8 +31,10 @@ class AddParticipant(Resource):
 
     @api.response(200, "Success")
     @api.response(400, "Validation Error")
-    @api.marshal_with(participant_model)
+    # @api.marshal_with(participant_model)
     def post(self, study_id: int):
+        if is_granted('viewer', study_id):
+            return "Access denied, you can not modify", 403
         study = Study.query.get(study_id)
         add_participant = Participant.from_data(request.json, study)
         db.session.add(add_participant)
@@ -47,6 +49,9 @@ class UpdateParticipant(Resource):
     @api.response(400, "Validation Error")
     @api.marshal_with(participant_model)
     def put(self, study_id, participant_id: int):
+        if is_granted('viewer', study_id):
+            return "Access denied, you can not modify", 403
+        is_granted('viewer', study_id)
         update_participant = Participant.query.get(participant_id)
         update_participant.update(request.json)
         db.session.commit()
@@ -55,6 +60,10 @@ class UpdateParticipant(Resource):
     @api.response(200, "Success")
     @api.response(400, "Validation Error")
     def delete(self, study_id, participant_id: int):
+        if is_granted('viewer', study_id):
+            return "Access denied, you can not modify", 403
+        is_granted('viewer', study_id)
+
         delete_participant = Participant.query.get(participant_id)
         db.session.delete(delete_participant)
         db.session.commit()
